Move benchmark execution onto TestRunner
